+++ src/controllers/UserRegisterController.php
@@ -6,18 +6,9 @@
 
 class UserRegisterController extends MainController
 {
-    function index()
+    function indexMain()
     {
-        $this->title = 'User Registation';
-
-        $parameters = func_num_args() >= 1 ? func_get_arg(0) : array();
-        
-        require ROOT_PATH . '/public/views/includes/header.php';
-        require ROOT_PATH . '/public/views/includes/nav.php';
-
         require ROOT_PATH . '/public/views/user-register/signin.php'; 
-
-        require ROOT_PATH . '/public/views/includes/footer.php';
     }
 }