+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 All notable changes to this publication will be documented in this file.
 
-## 0.1.0 - 2017-??-??
-Text
+## 0.1.0 - 2022-07-04
+MainController it's an abstract class now.
+Created a kind of a Factory Method for the controllers.
 
 ## 0.0.0 - 2022-07-04